move EntityTagSpec to a different place
