+++ rasa/core/policies/ted_policy.py
@@ -12,8 +12,7 @@ from typing import Any, List, Optional, Text, Dict, Tuple, Union, TYPE_CHECKING
 import rasa.utils.io as io_utils
 import rasa.core.actions.action
 from rasa.nlu.constants import TOKENS_NAMES
-from rasa.nlu.extractors.extractor import EntityExtractor
-from rasa.nlu.classifiers.diet_classifier import EntityTagSpec
+from rasa.nlu.extractors.extractor import EntityExtractor, EntityTagSpec
 from rasa.shared.core.domain import Domain
 from rasa.core.featurizers.tracker_featurizers import (
     TrackerFeaturizer,
